Renaming block so that directory structure matches block name. Change category to "Jetpack".
