+++ client/gutenberg/extensions/map/add-point/index.js
