+++ client/gutenberg/extensions/map/add-point/style.scss
