+++ client/gutenberg/extensions/map/editor.scss
@@ -1,6 +1,6 @@
 /** @format */
 
-.wp-block-jetpack-maps__delete-btn {
+.wp-block-jetpack-map__delete-btn {
 	padding: 0;
 	svg {
 		margin-right: 0.4em;
