+++ client/gutenberg/extensions/map/info-window/index.js
