+++ client/gutenberg/extensions/map/load-script/callback-handler.js
