+++ client/gutenberg/extensions/map/load-script/dom-operations.js
