+++ client/gutenberg/extensions/map/load-script/index.js
