+++ client/gutenberg/extensions/map/locations/style.scss
