+++ client/gutenberg/extensions/map/map-marker/style.scss
@@ -1,6 +1,6 @@
 /** @format */
 
-.wp-block-jetpack-maps-marker {
+.wp-block-jetpack-map-marker {
 	width: 32px;
 	height: 38px;
 	opacity: 0.9;
