+++ client/gutenberg/extensions/map/map-theme-picker/style.scss
