+++ client/gutenberg/extensions/map/mapbox-map-formatter/index.js
