+++ client/gutenberg/extensions/map/mapbox.scss
