+++ client/gutenberg/extensions/map/style.scss
@@ -1,7 +1,7 @@
 /** @format */
 
-.wp-block-jetpack-maps {
-	.wp-block-jetpack-maps__gm-container {
+.wp-block-jetpack-map {
+	.wp-block-jetpack-map__gm-container {
 		width: 100%;
 		overflow: hidden;
 		background: lightgray;
