+++ client/gutenberg/extensions/map/view.js
